Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -1,10 +1,7 @@
 #include <stdio.h>
 int main() {
-    int a;
+    double a;
     scanf("%lf",a);
     printf("You entered: %lf",a);
-
-
-    
     return 0;
 }
